mkutils: try loading extensions from package.path first
if this is not done they will always be loaded by the kpse loader. that
means during development we will always get the module installed on the
system, so we can't see or test changes to extensions without building
and installing the package. with this change we can set LUA_PATH
appropriately during development. (if in the root of the working tree
`LUA_PATH=../?.lua` should be sufficient